refact: União de Exception handlers.
