--- src/main/java/br/com/booksy/Booksy/exception/GlobalExceptionHandler.java
@@ -1,52 +1,0 @@
-package br.com.booksy.Booksy.exception;
-
-import br.com.booksy.Booksy.domain.dto.MessageDTO;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-    private static final Logger logger = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-
-    @ExceptionHandler(CommonException.class)
-    protected ResponseEntity<MessageDTO> handleCommonsException(CommonException exception) {
-        logger.error("[exception] {}", String.valueOf(exception));
-        return exception.getMessageError();
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseBody
-    public Map<String, String> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-
-        ex.getBindingResult().getAllErrors().forEach((error) -> {
-            String fieldName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-
-        return errors;
-    }
-
-    @ExceptionHandler(Exception.class)
-    protected ResponseEntity<MessageDTO> handleException(Exception ex) {
-        logger.error("Exception não tratada: ", ex);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(new MessageDTO("Exception não tratada: " + ex.toString(),
-                        "unichristus.global.handler.exception"));
-    }
-
-
-}
